pyproject.toml
[tool.poetry]
name = "m3u8"
version = "0.1.0"
description = ""
authors = ["neo <eyywqkgb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.0"}
gunicorn = "^20.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
aerich = "^0.7.1"
python-ffmpeg-video-streaming = "^0.1.15"
boto3 = "^1.26.129"
tzdata = "^2023.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-pagination = "^0.12.6"
sentry-sdk = {extras = ["fastapi"], version = "^1.29.2"}
authy = "^2.2.6"
qrcode = "^7.4.2"
aioredis = "^2.0.1"
oauthlib = "^3.2.2"
aiohttp = "^3.8.6"


[tool.poetry.group.dev.dependencies]
commitizen = "^2.42.1"
pre-commit = "^3.2.2"



[tool.aerich]
tortoise_orm = "migrations.conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
